Fix various file loading conversion warnings

+++ source/shared/loading/dotfileparser.cpp
@@ -419,7 +419,7 @@ bool DotFileParser::parse(const QUrl& url, IGraphModel* graphModel)
     it.onPositionChanged(
     [this, &fileSize](size_t position)
     {
-        setProgress(static_cast<int>((position * 100) / fileSize));
+        setProgress(static_cast<int>((position * 100) / static_cast<size_t>(fileSize)));
     });
 
     auto cancelledFn = [this] { return cancelled(); };
